refactor: align test names and packages with production layout
