+++ src/test/resources/records/mysql-region-delete.json
